Improve the Artifact detail Page
This code change includes improvements to the layout of a dedicated artifact page. We introduced segmented tabs of media, description and information. Media displays the media from the nft, description lays out the longer text that comes with an NFT, and information lays out the meta data.

Additionally, on the media page we began to utilize the underlying content of an NFT. In this case, we built functionality to display the text contents of a blog NFT. In the next code update, we will attempt to modularlize the content renderer into it's own component so we can have isolated functionality and begin tailoring to various media types.