Develop initial Audio Renderer component
In this code branch, we are attempting to develop the proof of concept for being able to support audio files. While testing this, we encountered an issue with the api that we utilize is not consistently pulling metadata.

For now, we are going to pause this implementation after this change and work on deploying the app instead of troubleshooting this issue only on local app deployment.